refactor(config-prefill): relocate pending-proposal to org-scoped router
Move GET pending-proposal from the v1 config router to the org-scoped agent
router (/api/:orgSlug/agents/:agentId/config/pending/:runId) so the client uses
the same base as every other config call — no /lobu/api/v1 special-casing. The
org-scoped placement also lets the query scope by organization_id (added a
cross-org 404 test). Test relocated to lobu/__tests__ over the real router; 6/6
green (agent proposal, watcher-args flatten + top-level agent_id, cross-agent
404, cross-org 404, missing-run 404, bad-id 400).

Bumps owletto pointer for the matching agentsUrl-base hook fix.
